+++ packages/dx-react-scheduler-demos/src/demo-sources/scheduler-editing/material-ui/copy-after-dragdrop.jsx
@@ -1,15 +1,17 @@
+/* eslint-disable react/no-unused-state */
 import * as React from 'react';
 import Paper from '@material-ui/core/Paper';
 import { ViewState, EditingState, IntegratedEditing } from '@devexpress/dx-react-scheduler';
 import {
   Scheduler,
-  DayView,
   Appointments,
-  AppointmentForm,
   AppointmentTooltip,
-  ConfirmationDialog,
+  WeekView,
+  DragDropProvider,
 } from '@devexpress/dx-react-scheduler-material-ui';
 import { appointments } from '../../../demo-data/appointments';
+
+const SHIFT_KEY = 16;
 
 export default class Demo extends React.PureComponent {
   constructor(props) {
@@ -17,21 +19,58 @@ export default class Demo extends React.PureComponent {
     this.state = {
       data: appointments,
       currentDate: '2018-06-27',
+      isShiftPressed: false,
     };
 
     this.commitChanges = this.commitChanges.bind(this);
+    this.onKeyDown = this.onKeyDown.bind(this);
+    this.onKeyUp = this.onKeyUp.bind(this);
+  }
+
+  componentDidMount() {
+    window.addEventListener('keydown', this.onKeyDown);
+    window.addEventListener('keyup', this.onKeyUp);
+  }
+
+  componentWillUnmount() {
+    window.removeEventListener('keydown');
+    window.removeEventListener('keyup');
+  }
+
+  onKeyDown(event) {
+    if (event.keyCode === SHIFT_KEY) {
+      this.setState({ isShiftPressed: true });
+    }
+  }
+
+  onKeyUp(event) {
+    if (event.keyCode === SHIFT_KEY) {
+      this.setState({ isShiftPressed: false });
+    }
   }
 
   commitChanges({ added, changed, deleted }) {
     this.setState((state) => {
       let { data } = state;
+      const { isShiftPressed } = this.state;
       if (added) {
         const startingAddedId = data.length > 0 ? data[data.length - 1].id + 1 : 0;
         data = [...data, { id: startingAddedId, ...added }];
       }
       if (changed) {
-        data = data.map(appointment => (
-          changed[appointment.id] ? { ...appointment, ...changed[appointment.id] } : appointment));
+        if (isShiftPressed) {
+          const changedAppointment = data.find(appointment => changed[appointment.id]);
+          const startingAddedId = data.length > 0 ? data[data.length - 1].id + 1 : 0;
+          data = [
+            ...data,
+            { ...changedAppointment, id: startingAddedId, ...changed[changedAppointment.id] },
+          ];
+        } else {
+          data = data.map(appointment => (
+            changed[appointment.id]
+              ? { ...appointment, ...changed[appointment.id] }
+              : appointment));
+        }
       }
       if (deleted !== undefined) {
         data = data.filter(appointment => appointment.id !== deleted);
@@ -57,19 +96,16 @@ export default class Demo extends React.PureComponent {
           />
           <IntegratedEditing />
 
-          <DayView
+          <WeekView
             startDayHour={9}
             endDayHour={17}
           />
           <Appointments />
-          <ConfirmationDialog
-            ignoreCancel
-          />
           <AppointmentTooltip
             showOpenButton
             showDeleteButton
           />
-          <AppointmentForm />
+          <DragDropProvider />
         </Scheduler>
       </Paper>
     );