+++ packages/dx-react-scheduler-demos/src/demo-sources/scheduler-featured-integrated-data-shaping/material-ui/demo.tsx
@@ -271,6 +271,7 @@ export default () => (
         timeTableCellComponent={TimeTableCell}
       />
       <DayView
+        displayName={'Three days'}
         startDayHour={9}
         endDayHour={17}
         intervalCount={3}