+++ packages/dx-react-scheduler-material-ui/src/plugins/edit-recurrence-menu.jsx
@@ -1,9 +1,9 @@
 import { withComponents } from '@devexpress/dx-react-core';
 import { EditRecurrenceMenu as EditRecurrenceMenuBase } from '@devexpress/dx-react-scheduler';
-import { Overlay } from '../templates/edit-recurrence-menu/overlay';
+import { Overlay } from '../templates/common/dialog/overlay';
 import { Layout } from '../templates/edit-recurrence-menu/layout';
-import { Button } from '../templates/edit-recurrence-menu/button';
-import { Container } from '../templates/edit-recurrence-menu/container';
+import { Button } from '../templates/common/dialog/button';
+import { OverlayContainer as Container } from '../templates/common/overlay-container';
 
 export const EditRecurrenceMenu = withComponents({
   Layout, Overlay, Button, Container,