+++ packages/dx-react-scheduler/src/index.ts
@@ -16,3 +16,4 @@ export * from './plugins/today-button';
 export * from './plugins/edit-recurrence-menu';
 export * from './plugins/integrated-editing';
 export * from './plugins/resources';
+export * from './plugins/confirmation-dialog';