+++ packages/dx-react-scheduler/src/types/editing/appointment-form.types.ts
@@ -336,5 +336,5 @@ export interface AppointmentFormProps {
 export type AppointmentFormState = {
   visible: boolean;
   appointmentData: AppointmentModel;
-  previousRule: string | undefined;
+  previousAppointment: AppointmentModel;
 };